+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # very-cool-table
 
+## 1.3.1
+
+### Patch Changes
+
+- 7c2d2dd: n-rows & n-cols property inheritance: true
+
+## 1.3.0
+
+### Minor Changes
+
+- 1e06dbe: log bugfix, css variables for table size
+
 ## 1.2.0
 
 ### Minor Changes